+++ keval/src/main/kotlin/com/notkamui/keval/AbstractSyntaxTree.kt
@@ -48,37 +48,6 @@ internal val KevalSymbolAnnotations: Set<ClassInfo> = ClassGraph()
         }.toSet()
 
 
-@KevalBinaryOperator(symbol = '+', precedence = 2, isLeftAssociative = true)
-fun KevalAdd(x: Double, y: Double): Double = x + y
-
-@KevalBinaryOperator(symbol = '-', precedence = 2, isLeftAssociative = true)
-fun KevalSub(x: Double, y: Double): Double = x - y
-
-@KevalBinaryOperator(symbol = '*', precedence = 3, isLeftAssociative = true)
-fun KevalMul(x: Double, y: Double): Double = x * y
-
-@KevalBinaryOperator(symbol = '/', precedence = 3, isLeftAssociative = true)
-fun KevalDiv(x: Double, y: Double): Double {
-    if (y == 0.0) throw KevalZeroDivisionException()
-    return x / y
-}
-
-@KevalBinaryOperator(symbol = '%', precedence = 3, isLeftAssociative = true)
-fun KevalMod(x: Double, y: Double): Double {
-    if (y == 0.0) throw KevalZeroDivisionException()
-    return y % x
-}
-
-@KevalBinaryOperator(symbol = '^', precedence = 4, isLeftAssociative = false)
-fun KevalPow(x: Double, y: Double): Double = x.pow(y)
-
-@KevalBinaryOperator(symbol = '(', precedence = 5, isLeftAssociative = true)
-fun KevalLPA(x: Double, y: Double): Double = 0.0
-
-@KevalBinaryOperator(symbol = ')', precedence = 5, isLeftAssociative = true)
-fun KevalRPA(x: Double, y: Double): Double = 0.0
-
-
 /**
  * Gets an operator by its symbol, currently checking only in this package.
  *
